+++ app/frontend/packs/application.js
@@ -25,6 +25,7 @@ import '../scripts/shared/itemEditForm'
 import '../scripts/shared/post'
 import '../scripts/shared/tooltip'
 import '../scripts/shared/mytask'
+import '../scripts/shared/githubLink'
 import 'scripts'
 import 'stylesheets'
 import '../stylesheets/index'
